fix(core): matching prefixed selectors without prefix configured
Fix the matching of prefixed selectors when no prefix is configured by
the built-in CSS and JavaScript language plugins.

This is achieved by splitting the Regular Expression for query selector
combiners into two: one for anything that is allowed before a selector,
and one for anything that is allowed after a selector; and using them in
both built-in language plugin.