fix(core): mangling attributes in HTML with comments and strings
Fix two bugs where the HTML language plugin's attribute mangling
capabilities fall short.

In the first bug attributes may be skipped if the element is in between
two HTML comments - see example below. This is due to the greedy
matching of HTML comments, causing it miss the end of one comment and
continue matching to the end of a later comment. Example (here "id"
would never be detected and mangled):

    <!-- foo --><div id="testing"><!-- bar -->

In the second bug attributes appearing after unquoted attribute values
would never be mangled. This is due to the part of the attributes regex
that matches individual attribute-value pairs which only knew about
valueless and quoted-value attrbutes. Example (here "id" may be detected
and mangled but "data-value" would never be):

    <div id=foo data-value=bar>