test(core): generate cases for HTML language plugin single value factory
Convert the test suite for the HTML language plugin's single value
attribute expression factory to be based on generated tests using
predefined value presets using the packages/testing new generating
functions.

This test suite iterates on the tests suites of other langauge plugin
expression factories with that are generated by converting the
`valueSets` into a function, allowing 1) for closure within the test
case to define variables local to the test case and 2) to delay value
generation in some cases until the test is started.