fix(core): mangling attribute values in HTML with comments and strings
Fix three bugs were the HTML language plugin's single-value attribute
capabilities fall short.

In the first bug attributes may be skipped if the element is in between
two HTML comments - see example below. This is due to the greedy
matching of HTML comments, causing it miss the end of one comment and
continue matching to the end of a later comment. Example (here the "id"
attribute value would never be detected and mangled):

    <!-- foo --><div id="testing"><!-- bar -->

In the second bug attributes appearing after unquoted attribute values
would never be mangled. This is due to the part of the attributes regex
that matches individual attribute-value pairs which only knew about
valueless and quoted-value attrbutes. Example (here the "id"  value may
be detected and mangled but the "data-value" value would never be):

    <div id=foo data-value=bar>

In the last bug the value of unquoted attribute-like strings in HTML
content would be mangled unintentionally - see example below. This is
because the (separate) regex for unqouted attribute values did not look
at all at whether or not the attribute is in a tag. Example (here the
"id" value would be detected unentionally; note that the spacing around
'id="foo"' is required to reproduce the bug):

    <div> id=foo </div>