chore: create test-runner script
Create a new script called "test.js" which provides the ability to run
tests with a bit more flexibility. It currently allows for easy
configuration of the type of test to run through a flag (currently
supported are "test" and "benchmark", but this could include e.g. "unit"
and "integration" in the future). And it currently allows to easily
choose what packages to run as "normal" arguments. E.g. to run the test
suites for packages/core and packages/cli you can use:

    $ npm run test -- core cli

The introduction of this script has several advantages:

- It removes the need for a separate command for each package.
- It allows for running two or more packages' test suite at once.
- New packages are automatically detected.
- Running test no longer relies on environment variables. Finally adding
  support for running tests on Windows.

Of course, the downside is an extra script that needs to be maintained.