fix(core): mangling attribute values in HTML with comments and strings
Fix a bug where the HTML language plugin's multi-value attribute may be
skipped if the element is in between two HTML comments - see example
below. This is due to the greedy matching of HTML comments, causing it
miss the end of one comment and continue matching to the end of a later
comment. Example (here the "class" attribute value would never be
detected and mangled):

    <!-- foo --><div class="foo bar"><!-- bar -->