chore: take care of compiling when running tests
Update the test.js script to take care of compiling the packages that
will be tested can be compiled, and that its local dependencies are
compiled.

This requried some general refactoring, mainly in order to avoid calling
`getPackagesToRun` twice. The result is a slightly cleaner base-level
flow in test.js. Additionally, the ESLint enforcement of JSDoc in
scripts has been disabled and some logging was added to the test script
for clarity when running tests - especially while compiling.