chore(benchmarking,testing): remove precompile cleaning
Remove the precompile command for both the benchmarking and testing
packages that cleans their folder. This speeds up compiling iff the
packages has previously been compiled.

This was originally needed because packages may be compiled into the
package's root. Now that all packages are compiled into a subdirectory,
cleaning the package is no longer necessary.