test(core): element-like attribute values with HTML attribute matcher
Add a test scenario to the test suite for the HTML language plugin's
attribute expression factory to test that the attributes of element-like
attribute values are not matched. I.e. if the value of "real" attributes
is set to an HTML element, e.g. `<div data-foo>`, the string "data-foo"
isn't matched.

The test scenario includes an additional attribute on the element to
ensure the attribute-value attribute is also not matched in the nested
expression. This implementation specific test is, clearly, also
generally applicable.