fix(scripts): exit with non-zero exit code when tests fail
Fix a bug with the test script where failed tests (regular tests, tests
with coverage, or benchmark tests) would never result in a non-zero exit
code. As a result, failing tests would not be detected by the CI.

The fix involves using `execFileSync` [1] instead of `spawnSync` [2],
which does throw an error if the created process exists with a non-zero
exit code (whereas `spawnSync` doesn't).

The call to `execFileSync` is abstracted internally and the error thrown
by it is caught, in which case the process is immediately exited with an
exit code of 1. This is because the error thrown by `execFileSync`
contains unnecessary detail, which is outputted if the error is not
caught.

--
1. https://nodejs.org/api/child_process.html#child_process_child_process_execfilesync_file_args_options
2. https://nodejs.org/api/child_process.html#child_process_child_process_spawnsync_command_args_options