fix(core): unintended mangling of HTML attributes
Fix unintended mangling of element-like attribute values in HTML by the
attribute mangling logic.

The change on line 17 prevents mangling directly in strings. That is it
prevents matching of, for example,

    "<div data-foo>"

in

    <div alt="<div data-foo>">

The change on line 23 prevents mangling in strings due to confusion
about quoteless attribute values. That is, it prevents matching of, for
example,

    data-foo>"

in

    <div alt="<div data-foo>">

with

    alt="<div

as quoteless attribute value. And lastly, the change on lines 32 to 38
prevents matching of, for example,

    data-foo

in

    alt="<div data-foo>" data-bar

after that was matched from

    <div alt="<div data-foo>" data-bar>