fix(scripts): running test coverage with subset of all packages
Fix a bug where running test coverage with only a subset of packages
would result in only those tests being run, but the coverage being
reported of all packages.

This was due to the nyc configuration not being properly updated when
`scripts/test.js` was introduced. It relied on outdated environment
variables.