perf(core): inserting embed identifiers into files
Significantly increase the perfromance of replacing embeds in file
content strings by replacing string concatenation with a "string
builder" - which in practice translates to `push`ing string parts into
an array and `join`ing them together in the end.

Anecdotally, this reduces the time to insert 3852 identifiers in a 5MB
HTML file from ~20 seconds to practically 0 seconds.