fix(core): mangling CSS values in CSS function
Fix support for mangling CSS values in CSS functions. This is achieved
by 1) allowing the following characters before the value:

    "(", "+", "-", "*", "/", ","

Where the first and last character are standard characters in any
function, and the other characters are arithmetic operators for the
`calc()` function.

And by 2) allowing the following characters after the value:

    ",", "+", "-", "*", "/", ")"

where, again, the first and last values are standard characters in any
function and the others are arithmetic operators for the `calc()`
function.