chore(deps): add Stryker core and typescript-checker
Add Stryker [1], a mutation testing framework [2] by adding Stryker core
and Styker TypeScript checker as dependencies to the project. The former
is the framework's core logic and the latter is an additional tool that
levarages TypeScript type annotation to avoid mutating in ways that
contradict the typing (speeding up testing).

The goal of adding mutation testing is to uncover holes in the unit test
suites for the various projects.

--
1. https://stryker-mutator.io
2. https://en.wikipedia.org/wiki/Mutation_testing