chore(scripts): add mutation testing to test runner
Add support for running mutation tests through Stryker through the test
runner script (`scripts/test.js`) using the "--mutation" flag. The test
runner will, as usual, take care of selecting packages and running the
right command. A shorthand for running mutation tests with the script is
added to the package manifest under `npm run test:mutation`.

To support running mutation tests using Stryker, a Stryker configuration
file is added. It is configured for performance and will only mutate the
package(s) selected through the runner. Additionally, it is configured
to work optimally with TypeScript (avoid type check errors and disabling
mutations that would break the type system). Lastly, the reports and
temp directory are configured in line with this project's defaults. This
configuration file was also added to ESLint.

Note that `scripts/test.js` now also looks at the `process.env.
TEST_PACKAGES` value when determining what packages to run. This is
REQUIRED for mutation testing. Stryker re-invokes the test command to
see if mutants are killed, but it won't include the names of the
packages to test when doing so (resulting in all tests being run).
However, Stryker does keep the `TEST_PACKAGES` environment variable
and passes it on to the test command.