benchmark(core): getEmbeds and reEmbed
Benchmark the getEmbeds and reEmbed functions of the core package. The
benchmarks are designed to test the perfromance improvements of [1] and
[2]. The number of embeds determines the difficulty of the benchmark for
`getEmbeds` while the size of the file determines the difficulty of the
benchmark for `reEmbed`.

These benchmarks were also ran against the `main` branch (commit [3])
which resulted in the following median runtimes:

- `getEmbeds`: ~48ms
- `reEmbed`: ~15ms

--
1. 4a40b498446095c5dceaec4c6f332017546c4427
2. 0fa5e68ae308a3ac4f036593daeede6ecd026b4a
3. afff19a0fcbfbddbca8b3a3f52d16d29da90eaf5