chore(benchmarking): kill mutation test mutants
Improve the test suite and source the benchmarking package to kill all
mutants identified by mutation testing using Stryker.

Stryker identified three surviving mutants in the benchmarking package.
The first mutant is in budget.ts and relates to the fact that none of
the tests had more than one CPU, as a result to `getCpuSpeedInMHz`
function was not properly covered and could be changed without failing
tests.

The second mutant is in the runner.ts and relates to the fact that none
of the tests verify the duration of a single run is computed correctly.
Unfortunately this computation is tightly coupled with stats.ts and
therefore a bit hard to tests. Nevertheless a basic test case has been
added to kill the mutant.

Lastly, another mutant in runner.ts. This was actually a harmless mutant
but it was killed nonetheless. In this case the return value of `NOOP`
(in `getSetupFn`) could be changed from `null` to `undefined`. This is
fine in practice, but to prevent to mutant from being reported the
return value is replaced by an empty function.