test(core): other strings in JS language plugins query selector tests
Introduce the presence of other strings in the JavaScript language
plugin's test suite for the CSS query selector expression factory.
This addition exposes the bug where sometimes CSS query selectors would
go undetected deu to the expressions mismatching quotes.

To support this, a "beforeStatement" key was introduced for the
JavaScript statement values type, allowing for the addition of JS
before the statement being tested. Which is required for the test cases
being added in this commit. The builders.ts for the JS expression
factory tests is updated for this addition, and it's test suite
verifies the "beforeStatement" is used correctly.