fix(core): mangling query selectors in JavaScript
Fix problems related to mangling query selectors in JavaScript due to
improper quote matching in the JavaScript built-in language plugin.
This is fixed by always matching quotes of pairs iff a quote is
encountered. If the string in between does not contain the expression
being searched for the match is simply thrown away.

The related test suite had to be updated for performance reasons.
Uninteresting test scenarios, `beforeLeftHand` and `afterLeftHand` have
been removed as they're unlikely to affect the language plugin in any
way that, e.g., `beforeRightHand` and `afterRightHand` do. Moreover, the
value presets `beforeStatement` and `afterStatemen` have been reduced.
Again, uninteresting values, such as whitespace, were removed.