benchmark(core): adjust embeds.ts budgets
Increase the embeds.js::getEmbeds budget to 3 milliseconds. Before the
addition of the embed id prefix the function barely achieved the
benchmark goal (making it potentially flaky) and with the addition of
the benchmark id prefix the benchmark budget is just a little bit too
low. Interestingly, shortening the prefix ensures the benchmark goal is
achieved. However, I think a sufficiently long (and hence more unique)
prefix is more important than a less than 1ms performance gain.

Additionally, the embeds.ts::reEmbed function's benchmark budget was
lowered to 1ms  as analysis revealed its performance is currently below
1ms, to encourage that level of performance when the function is changed
in the future the benchmark budget was changed accordingly.