refactor(core,types): extract WebMangler types into a dedicated package
Extract types for WebMangler into a dedicate package. This will prevent
circular dependencies where the core package wants to import other
packages, which in turn need the type definitions previously defined in
the core package.

The starting version of this package will equal to the current version
of the core package - this allows for keeping the at-since and
at-version metadata in the type documentation.