refactor(core,lang-utils): extract language plugin utils into a package
Extract the language pluign utilities of the WebMangler core into a
separate package. This will prevent
circular dependencies where the core package wants to import other
packages, which in turn need the language utilities provided by the
core.

The starting version of this package will equal to the current version
of the core package - this allows for keeping the at-since and
at-version metadata in the type documentation.