refactor(core,lang-css): extract CSS language plugin into a package
Extract the CSS language plugin of the WebMangler core into a
separate package. This ensures separation of concerns, will make it
easier to swap standard language plugins later, and reduces the number
of tests that need to run per package.

The starting version of this package will equal to the current version
of the core package - this allows for keeping the at-since and
at-version metadata in the type documentation.

Of particular note are the changes to the CSS language plugin's
benchmarking tests. Before, they used the WebMangler core's engine. This
dependency on the WebMangler core has been removed by instead measuring
the performance of the `findAll` method (in the future `replaceAll`
could be benchmarked separately, but that is beyond the scope of this
commit). Not only does this prevent a circular dependency between the
CSS language plugin and the WebMangler core, it also focusses the
benchmarking on relevant code as it can now ignore the time other parts
of the engine take - as evidenced by the reduced budgets.

Do note that the conversion to arrays **inside** the benchmarked
function is crucial. This is because `findAll` is (/may be) a generator,
which won't do any work if simply called.