chore: add .nvmrc with Node v15.14.0 (and NPM v7.7.6)
Add an `.nvmrc` file (for NVM, Node Version Manager [1]) with v15.14.0,
allowing contributors to run

  $ nvm use

at the root of the project and use Node v15.14.0. This version of Node
was chosen as it comes with NPM v7 and as opposed to v16.x is a bit more
mature. However, v15.x has no official LTS plans, and so this project
should switch to v16.x once that is mature (i.e. infrequent minor
updates).

The switch from NPM v6 (or lower) to v7 comes with improvements to the
`package-lock.json` format. These improvements are included for the root
`package-lock.json` of the project in this commit.