test(lang-js): add comment with comment characters test cases
Update the `buildJsInlineComment` function to produce more than one
comment. In particular, it produces one version of the comment with "*"
inserted before the comment, and one with "/" inserted before the
comment.

This increases the number of test cases for tests using this function
dramatically, hence relates tests have been updated to reduce the number
of tests.