refactor(core): extract mangler plugin utilities into a separate package
Extract the mangler plugin utilities of the WebMangler core into a
separate package. This will prevent circular dependencies where the core
package wants to import other packages, which in turn need the mangler
utilities provided by the core.

The starting version of this package will equal the current version of
the core package - this allows for keeping the at-since and at-version
metadata in the documentation.