chore(deps): replace eslint-plugin-md by eslint-plugin-markdown
Remove the plugin eslint-plugin-md [1] because (a) it has caused
dependency before, and (b) it is causes dependency issues with remark
updates. As [1] seems to have been abandoned, this adds the linting of
snippets in MarkDown using eslint-plugin-markdown [2]. However, this
plugin does not lint MarkDown itself, so a substitute for that has to
be added.

To make this work the ESLint configuration had to be updated. Notably,
[2] was having some errors due to the at-typescript-eslint/parser so
the config is updated so that only *.ts files use that parser.

--
1. https://www.npmjs.com/package/eslint-plugin-md
2. https://www.npmjs.com/package/eslint-plugin-markdown