test(lang-html): add suite testing common patterns directly
Add a test suite for the common patterns of the langauge-html package.
This should increase confidence in both these patterns, and the larger
patterns that use them.

Note that the negative test suite for `patterns.attributes` is skipped.
This is because the pattern in question currently matches the empty
string, as a result it will `.test` true against every string.