chore(scripts): add script to generate workflow files for packages
Add a script that can generate a workflow file for a package of the
project. This includes setting path filters for when the workflow should
be run; only running tests if the package has tests; only running
benchmarks if the package has benchmarks.

The paths helper script got additional helper functions to resolve
paths to support the above script.