ci: add workflow file for each package
Add a workflow file dedicated to compiling as well as running the tests
and benchmarks for that package, using package filters to ensure the
workflow is only run when the package has changed.

Note that these workflows are automatically generated and should not be
edited manually.

The old push-checks workflow has been removed in favor of these new
workflows. The lint job has been moved to the Code Analysis workflow.