test(lang-js): add suite testing common patterns directly
Add a test suite for the common patterns of the langauge-js package.
This should increase confidence in both these patterns, and the larger
patterns that use them.

The regular expression for line comments was updated to match empty line
comments. This does not affect the language plugin (based on the test
suite) but does line-up the expression with its documentation