ci: construct build, test, and benchmark matrices dynamically
Generate the execution matrix for building, testing, and benchmarking
packages dynamically. Still using dorny/paths-filter[1], this introduces
a new job (preceding build, test, and benchmark) that determines which
packages were changed and thus should be verified in the CI.

This way, jobs are only created for packages that were actually changed,
rather then skipping steps in the jobs of packages that were not
changed.

It should be noted that in the current implementation, new packages
still have to be added to the workflow file manually. Namely, as a
filter in the 'changes' job.

--
1. https://github.com/dorny/paths-filter