ci: generate dorny/paths-filter filters based on the packages folder
Replace the "hardcoded" list of filters to determine which packages
were "changed" by a dynamically generated string. This string is
generated by a new script (scripts/ci/get-filters.js) that looks at the
folders inside the packages/ folder and creates a filter for each. Note,
the string replacements in the workflow step with id "filters" resolve
potential problems with special characters (specifically "%" and
newlines) when it comes to outputting it in GitHub Actions.