test(lang-css): single value attribute in a media query
Add a collection of tests for the single value attribute expression
factory to see if matches single value attribute selectors inside of a
media query.

The `CssRulesetKey` test type was updated to include "beforeRuleset" and
"afterRuleset", used here for putting a ruleset inside of a media query.
Test helpers that use this type have been updated accordingly.