chore: upgrade Node.js version for development to v16
Change the Node.js version for development from v15 to v16. v15 was an
odd choice because it's not an LTS release, so v16 makes more sense. It
is pinned to 16.8.0 for increased stability.

Additionally, update the "name" of the step that sets up NodeJS in the
CI so that it doesn't have to be changed when the Node.js version is
changed in the future.