chore: update node version range, add npm version range
Update the node version range in the root package.json to encourage
developers to use one of the supported LTS release of Node.js.

Also add an `npm` version range in the root package.json to match the
versions required to install dependencies.

The lockfile is updated in accordance with these changes.