ci(deps): pin all Actions at the latest version
Pin all Actions used in GitHub Actions to the latest release version.
From a security perspective this ensures that the CI does not
automatically use the latest, potentially compromised, version of the
Action but instead Action updates can be provided by Renovate (or
similar) and reviewed. Clearly this is not a guarantee, but it's a best
effort at improving the security of the CI.