chore: enable TypeScript member delimiter style in ESLint
Enable the member delimiter rule [1] for the TypeScript ESLint plugin
with a warning. A warning was chosen as there are currently various
violations  that should be fixed when the respective file (or in some
cases package) is changed.

This rule should ensure the type and interface definitions are
consistently
formatted w.r.t. delimiters, always with a `;` and never omitted.

--
1. https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/member-delimiter-style.md