chore: initialize npm workspaces
Initialize npm workspaces [1] in the project. This allows for a couple
of things, most notably having to run `npm install` only from the root
of the project, being able to run scripts of package from the root, and
not having to install TypeScript seperately for each project.

The package manifest of the CLI package was updated to create the binary
executable file before installing dependencies. This is currently
required as installing the dependencies of this package as a workspace
results in an error if that file doesn't exist (npm tries to `chmod` the
executable file as part of the install command through workspaces, for
some reason).

--
1. https://docs.npmjs.com/cli/v7/using-npm/workspaces