ci: update dependency installation and compile job
Update how dependencies are installed in the CI, as well as how the
compile job works in-line with the adoption of npm workspaces.

The npm cache in the "Determine jobs" job was removed as the job does
not install any dependencies, leading to potential errors when the job
is executed.