chore(types): export MangleExpression option type definitions
Export the newly added type defintions for `MangleExpression` options in
the types package.

The type package's root barrel file is also refactored to omit the
`export x from y` syntax in favor of import-then-export. This
will simplify the diff when the location of an exported variable changes
- clearly showing that the exported values of the packages don't change.

The CHANGELOG of the package is also updated to include this change.