ci(scripts): change filters to trigger on any change in a package
Update the `get-filters.js` script to return filters that match any file
in a package, rather than just files in the src directory. This ensures
the CI for a package is run whenever there's a change to e.g. the
package manifest. Unfortunately it also means it is triggered when
changing the package's documentation (which would't affect tests etc.).
At least in this point in time, this trade-off is worth it as running
tests etc. when the package manifest changes is more improtant than
not having to wait for CI when documentation is changed.