chore(scripts): create script to run linters
Create a script to run linters. Currently, it wil run eslint, optionally
with the "--fix" flag on files that have been changed w.r.t. to the
`main` branch. If the "--all" flag is provided to this scirpt it will
instead lint all files.

The push-checks workflow's lint job was updated to fetch the main
branch explicitly. This ensures it can find the files changed since
branching off of `main`.