refactor(testing): update import-export style
Update the importing and exporting in package testing such that imports
and exports all appear on individual lines. This helps increase identify
what is and what isn't exported (to avoid accidentally removing
something that should be exported for backwards compatability) as well
as simplify the diff when exports/imports change (in the spirit of "a
line of code should do only one thing").