refactor(testing): use dependency injection for mock classes
Refactor the mock classes provided by package testing to use dependency
injection so that these classes can be unit tested with zero
dependencies. Unit tests for these classes are also added in this
commit.

This involves two major changes. First is the refactoring of the mock
classes' modules to allow for dependency injection. Second is the
replacement of the common function `getStubOrDefault` by a simpler
logical or inside the classes (with the mocks creating stubs themselves
instead). This was chosen over injecting `getStubOrDefault` as it
makes dependency injection more general and easier to manage as well as
test.

This commit also includes a minor bug fix for the MangleExpressionMock,
its fallback `replaceAll` stub now returns a string (as per the
interface it's supposed to implement).