chore: initialize markdownlint
Initialize markdownlint with a configuration file and a temporary
`npm run-script` command to test the configuration. The configuration
is mostly standard, with only rule MD024 [1] configured to allow for
identical headers if-and-only-if the headers are not siblings. Identical
headers are currently used at lower levels for common section (such as
an "Examples" section).

The ESLint configuration is updated to lint this new configuration file.
All dot-files with the `.yml` extension are now explicity not ignored,
this is needed because ESLint ignores dot files by default.

--
1. https://github.com/DavidAnson/markdownlint/blob/f3f8d635d58387cc78d0a66e394a0ad925b3301f/doc/Rules.md#md024