chore: rename "benchmark" script to "test:performance"
Rename for increased clarity. Benchmark is a term that is somewhat
ambiguous here. Most problematic is the fact that "benchmark" doesn't
necessarily imply a test. The perhaps more straightforward change to
"test:benchmark" was not chosen because "benchmark" has a more common
meaning in terms of comparing performance, whereas here we are testing
by requiring function to run (on average) faster than a certain limit.
Hence, "test:performance" was chosen to more clearly communicate what
the script does, it tests the performance of the code.