feat(scripts): add get-filters for mutation testing
Add support for a new argument, "mutation", to get a dorny/paths-filter
filters string for the packages that are ready to be mutation tested.
Ultimately, this would be all (tested) packages. Currently, it's limited
to a hardcoded list of packages that have mutation tested and are ready
for the level of scrutany mutation testing brings.