chore(language-css): use options types from package types
Remove the local definitions of the type options and use those from the
types package instead. Update the package manifest to require the latest
version of the types package.

Also remove the no-longer-needed lockfile and ignore it explicity in
.gitignore. (It is no longer needed due to the use of npm workspaces.)

Also apply some stylistic changes to files changed in this commit.