ci: replace workflow path filters by dynamic path filters
Remove the path filters from the workflow as they duplicate the
`get-filters.js` script (reverting [1]) and update `get-filters.js` to
return filters that include the root `package-lock.json` for every
package, ensuring all packages are tested when dependencies change (a
drawback of npm workspaces being that the per-package lockfile cannot be
used to limit the number of jobs), as well as for the workflow file
itself, which is to ensure that changes to it work as expected.

The script file is also formatted in accordance with ESLint.

--
1. 2c4ee7ca0838627a5298026d99843fef0da0cb6e