config: define test structure for performance tests
Update the test configuration for performance tests - currently defined
as a hybrid of the existing setup and the new setup. The new setup will
consist of a "performance" folder inside "__tests__" folders with files
named "xyz.test.ts", significying that they are perforamnce **tests**.

By adopting the `.test.ts` suffix for performance tests causes some
complications with the existing hybrid definitions for all tests vs.
unit and performance tests, hence the `testTypeDir` for all tests is now
explicitly defined (instead of implicitly derived from the fact that
performance tests use the ".bench.ts" suffix).

Note: a `.perf.ts` suffix was considered, but to promote the idea that
these are performance **tests** the `.test.ts` suffix was chosen
instead.