config: enable ESLint for .cjs configuration files
Update the ESLint configuration so that it lints CommonJS confugiration
files. Update the ESLint configuration to accomodate this in an expected
manner, in particular adding `__dirname` as a readonly global and not
requiring JSDoc in these configuration files.

The configuration file that did not yet comply with the existing format
have been reformatted to comply with the existing ESLint configuration.