fix(scripts): update `hasFiles` filters
Update the `hasFiles` filter of the `get-filters.js` script to detect
tests and performance tests correctly with the upcoming naming
conventions for performance tests. Notably, the script now explicitly
separates expressions for legecy and future naming; and the script will
now not mistake performance tests (with the `.test.ts` suffix) for
"normal" tests.