ci: upload mutation report as artifact
Use the actions/upload-artifact action [1] to upload the mutation report
when running mutation tests for a package. Each report is identified by
the package name in the artifact name. Additionally, the commit hash is
included in the artefact to be able to easily determine the commit for
which the report is.

--
1. https://github.com/actions/upload-artifact