feat(mangler-utils): add `ALL_ALPHANUMERIC_CHARS` character preset
Add a new character preset consisting of all letters (uppercase and
lowercase) as well as all numeric characters.

Deprecate the `ALL_CHARS` character preset as it's somewhat dangerous to
use in practice since on the one hand it would always need to include
all characters (defined in `CharSet`) but on the other hand adding new
characters to it should be considered a breaking change.

Additionally fix a typo in the JSDoc for the `ALL_LETTER_CHARS` preset
and improve the clarity of the `ALL_NUMBER_CHARS` preset.