chore(benchmarking): split unit and integration tests in folders
Split the unit and integration tests for the benchmarking package.
Notably, the `index.test.ts` suite is split into two separate files (a
unit suite checking exports and an integration suite checking behaviour)
and the `stats.test.ts` now uses the (new) `TestScenarios` type from the
testing package - this is also the reason for the change to the
TypeScript configuration for the benchmarking package.