refactor(scripts): update linter script for flexibility
Rework the linter script to allow for A) linting a specific language,
and B) linting multiple individual languages in one go. This provides
flexibility to run linters only for a subset of languages that can be
linted as well as the flexibility to add more then one linter for a
language.