chore(scripts): run markdownlint with lint.js
Update the lint.js script to run markdownlint in a way that is
approximately equivalent to the (now removed) npm script `lint:md`. The
`--ignore-path` is still set to `.gitignore` to prevent linting VCS
ignored files, but - given how lint.js works - the default way
markdownlint is run now is only on changed files as opposed to all files
with `lint:md`; unless the `--all` flag is in which case it's the same
as `lint:md`. Also, the `--dot` flag of markdownlint is now used, this
is to ensure that MarkDown files nested in dot-folders (e.g. ".github")
are linted as well. Lastly, linting MarkDown with lint.js differs from
`lint:md` in that it also (always) lints code snippets in MarkDown using
ESLint.