refactor(benchmarking): runner logic for testability
Refactor the runner logic (`src/runner.ts`) such that it behaves
identically but takes dependencies - in this case the "perf_hooks"
module [1] and internal depencencies. This is hidden from the end-user
of the package by instantiating the functions with dependencies in the
runner's barrel file.

The type definitions for the runner have been moved to a local
`types.ts` file.

--
1. https://nodejs.org/api/perf_hooks.html