fix(scripts): linting with deleted files
Fix a bug in the `lint.js` script that resulted in no files being linted
due to an error when a previously changed file is deleted but not yet
committed. In this scenario the script would attempt to lint the removed
file, (potentially, depending on the
linter) causing the linter to fail.

The fix involves obtaining the git status information from
git-changed-files [1] and filtering based on the status of uncommitted
files.

--
1. https://www.npmjs.com/package/git-changed-files