refactor(language-utils): object-as-parameter for nested group expr
Update the `NestedGroupMangleExpression` constructor to accept an object
as parameter and deprecate the existing constructor construct. The test
suite for the class now runs the same tests for both version of the
constructor. The documentation for the constructor is currently focussed
on the old constructor and should be updated when the deprecated version
is removed.

Additionally, refactors the test for `NestedGroupMangleExpression` to
use the new webmangler/testing's `TestScenarios` type (over the
deprecated `TestScenario` type).