refactor(benchmarking): stats logic for testability
Refactor the runner logic (`src/stats.ts`) such that it behaves
identically but takes dependencies - in this case only internal
depencencies. This is hidden from the end-user of the package by
instantiating the functions with dependencies in the stats' barrel
file.

The type definitions for the stats have been moved to a local
`types.ts` file.