ci: enable continuous mutation testing for package benchmarking
To prevent errors when running mutation tests for this package without
cleaning the compiled code first, the Stryker configuration is updated
as well. Namely, due to renaming of files, the `build/` folder of the
package may contain compiled versions of now-removed files. This would
cause an error in Stryker's typechecker. However, the typechecker
doesn't have to typecheck compiled code, so Stryker now uses the
`tsconfig.build.json` configuration file for typechecking. This file
ensures the compiled directories (and some others) are ignored when
typechecking.