feat(scripts): only run tests for changed packages by default
Update the test.js script to run only the tests ("normal", unit,
integration, performance, mutation) for VCS changed packages _by
default_. That is, when not specifying what package(s) to test.

Additionally, add support for an `--all` flag which can be used to
invoke the old behaviour where all packages are tested in case that is
desired.