test(language-utils): improve mangle-expression helpers suites
Improve the test suites for the mangler-expression helpers of the
language utilities package based on mutation tests results. These
categories of tests have been added for both helpers:

- Tests verifying the default value of the `caseSensitive` option.
- Tests verifying newlines in patterns are handled correctly.
- Tests verifying casing is preserved when the `caseSensitive` option
  is set to `true`.

For the `NestedGroupMangleExpression` the test suite is also updated to
verify that an empty replacement map is handled correctly by
`replaceAll` *and* to include a test verifying that matches that do not
match on the specified group are handled correctly.

The tests were not improved to prevent timeouts due to mutants, nor were
they updated to catch mutants in deprecated code.